Changed Alternatives usage for Debian
The way that `vim-formula` currently handles the alternative settings
for the `/usr/bin/editor` symlink on Debian based machines is incorrect,
this was brought up in saltstack-formulas/vim-formula#29.

I tested this on Debian, Ubuntu, and CentOS (the `editor` symlink does
not exist by default on this OS) and it seems to work as
expected now.