fix(vim/init.sls): fix full vim installation on suse
On Suse vim is split to 3 packages:
  - vim
  - vim-data-common (common data files for vim & gvim)
  - vim-data (data files)
vim-data-common is required, but vim-data is only recommended.
If solver.onlyRequires is enabled in Zypper (e.g. in JeOS),
only core vim is installed (without runtime *.vim files).
This commit allows to specify a list of packages for installation
and installs both vim and vim-data by default on Suse os_family.